requirements.txt
# 企业微信LLM交互服务端依赖

# 核心依赖
python-dotenv>=1.0.0
openai>=1.0.0
wechatpy>=1.8.18
requests>=2.28.0

# 工具模块依赖（复用local_client的工具）
# 这些工具模块已经在项目中，不需要额外安装
ddgs>=8.0.0

# 开发依赖（可选）
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0